Updated MTO
git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@892 b725f6ce-2418-43e8-b868-1e0ad3d2b257

+++ src/cpp/mto/manager/peercollection.cpp
@@ -75,7 +75,7 @@ PeerConnectionHandler *PeerCollection::create(ClientSocket *sock, std::vector<Mt
 
 void PeerCollection::update(PeerConnectionHandler *receiver, const Header &h)
 {
-    MtoHello hello(h.src.port, h.dst.port, h.length);
+    MtoHello hello(h.src.port, h.dst.port, (uint16_t)h.length);
     if(!insert(receiver, hello))
         // if hello is invalid or already known
         return;
